% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC.R
\name{gsea_prot_prob_assign}
\alias{gsea_prot_prob_assign}
\title{This assigns probabilities to individual proteins based on GSEA}
\usage{
gsea_prot_prob_assign(cplx, p)
}
\arguments{
\item{cplx}{complex vector}

\item{p}{is the complex-based GSEA probability, typically should be the 1 minus the result from certain sample of repgsea}
}
\value{
output The vector of GSEA probabilities for the proteins of a sample
}
\description{
This assigns probabilities to individual proteins based on GSEA
}
