% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC.R
\name{PROTREC_protprob}
\alias{PROTREC_protprob}
\title{This assigns probabilities to individual proteins based on PROTREC}
\usage{
PROTREC_protprob(cplx, p, prot, fdr)
}
\arguments{
\item{cplx}{complex vector}

\item{p}{is the complex-based PROTREC probability, typically should be the 1 minus the result from certain sample of PROTREC_cplx_prob}

\item{prot}{is the list of observed proteins in the screen}

\item{fdr}{is the false discovery rate of the screen}
}
\value{
The vector of PROTREC probabilities for the proteins of a sample
}
\description{
This assigns probabilities to individual proteins based on PROTREC
}
