% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC.R
\docType{data}
\name{RC}
\alias{RC}
\title{This is the RC protein data included in my package}
\description{
This is the RC protein data included in my package
}
\references{
\url{http://www.nature.com/nm/journal/v21/n4/full/nm.3807.html}
}
